+++ lib/deepar_flutter_plus.dart
@@ -1,5 +1,5 @@
 library deepar_flutter_plus;
 
-export 'src/deep_ar_controller.dart';
-export 'src/deep_ar_preview.dart';
+export 'src/deep_ar_controller_plus.dart';
+export 'src/deep_ar_preview_plus.dart';
 export 'src/resolution_preset.dart';